+++ apps/bbsync/save.py
@@ -1,6 +1,7 @@
 from collectors.bzimport.collectors import BugzillaConnector
 from osidb.models import Flaw
 
+from .exceptions import UnsaveableFlawError
 from .query import BugzillaQueryBuilder
 
 
@@ -10,19 +11,15 @@ class BugzillaSaver(BugzillaConnector):
     flaw validity is assumed and not checked
     """
 
-    class UnsaveableFlaw(Exception):
-        """
-        error caused by attempt to save a flaw which cannot be saved
-        either by its nature or due to the current saver capabilities
-        """
-
-        pass
-
-    def __init__(self, flaw):
+    def __init__(self, flaw, bz_api_key):
         """
         init stuff
         """
         self.flaw = flaw
+        # substitute the default service Bugzilla API key
+        # so the resulting Bugzilla audit log corresponds
+        # to the acutal user requesting the operation
+        self._bz_api_key = bz_api_key
 
     def save(self):
         """
@@ -48,7 +45,7 @@ class BugzillaSaver(BugzillaConnector):
         # TODO flaws with multiple CVEs are non-trivial to update
         # and save back to Bugzilla so let us restrict this for now
         if Flaw.objects.filter(meta_attr__bz_id=self.flaw.bz_id).count() > 1:
-            raise self.UnsaveableFlaw(
+            raise UnsaveableFlawError(
                 "Unable to save a flaw with multiple CVEs to Bugzilla "
                 "due to an ambigous N to 1 OSIDB to Buzilla flaw mapping"
             )