+++ osidb/serializer.py
@@ -13,6 +13,7 @@ from django.core.exceptions import ValidationError
 from drf_spectacular.utils import extend_schema_field, extend_schema_serializer
 from rest_framework import serializers
 
+from apps.bbsync.mixins import BugzillaSyncMixin
 from apps.osim.serializers import WorkflowModelSerializer
 
 from .core import generate_acls
@@ -366,6 +367,35 @@ class ACLMixinSerializer(serializers.ModelSerializer):
         validated_data = self.embargoed2acls(validated_data)
         return super().update(instance, validated_data)
 
+    def validate(self, data):
+        """
+        validate that the current user is member of all LDAP groups (s)he provides
+        with the access so there is no access expansion beyond the user permissions
+        """
+        data = super().validate(data)
+
+        # get the resulting ACLs from the embargo status
+        embargoed = self.context["request"].data.get("embargoed")
+        acl_read = (
+            settings.EMBARGO_READ_GROUP if embargoed else settings.PUBLIC_READ_GROUPS
+        )
+        acl_write = (
+            settings.EMBARGO_WRITE_GROUP if embargoed else settings.PUBLIC_WRITE_GROUP
+        )
+        acl_read, acl_write = ensure_list(acl_read), ensure_list(acl_write)
+
+        acls = [group.name for group in self.context["request"].user.groups.all()]
+        for acl in acl_read + acl_write:
+            # this is a temporary safeguard with a very simple philosophy that one cannot
+            # give access to something (s)he does not have access to but possibly in the future
+            # we will want some more clever handling like ProdSec can grant anything etc.
+            if acl not in acls:
+                raise serializers.ValidationError(
+                    f"Cannot provide access for the LDAP group without being a member: {acl}"
+                )
+
+        return data
+
 
 class MetaSerializer(ACLMixinSerializer, TrackingMixinSerializer):
     """FlawMeta serializer"""
@@ -401,8 +431,44 @@ class CommentSerializer(TrackingMixinSerializer):
         ] + TrackingMixinSerializer.Meta.fields
 
 
+class BugzillaSyncMixinSerializer(serializers.ModelSerializer):
+    """
+    serializer mixin class implementing special handling of the models
+    which need to perform Bugzilla sync as part of the save procedure
+    """
+
+    bz_api_key = serializers.CharField(allow_null=False, required=True, write_only=True)
+
+    def create(self, validated_data):
+        """
+        perform the ordinary instance create
+        with providing BZ API key while saving
+        """
+        bz_api_key = validated_data.pop("bz_api_key")
+        instance = self.Meta.model(**validated_data)
+        instance.save(bz_api_key=bz_api_key)
+        return instance
+
+    def update(self, instance, validated_data):
+        """
+        perform the ordinary instance update
+        with providing BZ API key while saving
+        """
+        bz_api_key = validated_data.pop("bz_api_key")
+        for attr, value in validated_data.items():
+            setattr(instance, attr, value)
+        instance.save(bz_api_key=bz_api_key)
+        return instance
+
+    class Meta:
+        model = BugzillaSyncMixin
+        fields = ["bz_api_key"]
+        abstract = True
+
+
 class AffectSerializer(
     ACLMixinSerializer,
+    BugzillaSyncMixinSerializer,
     TrackingMixinSerializer,
     IncludeExcludeFieldsMixin,
     IncludeMetaAttrMixin,
@@ -426,8 +492,6 @@ class AffectSerializer(
         "ps_component",
         "ps_module",
         "resolution",
-        # Internal data
-        # "acl_labels",
     )
 
     trackers = serializers.SerializerMethodField()
@@ -479,6 +543,7 @@ class AffectSerializer(
                 "delegated_resolution",
             ]
             + ACLMixinSerializer.Meta.fields
+            + BugzillaSyncMixinSerializer.Meta.fields
             + TrackingMixinSerializer.Meta.fields
         )
 
@@ -517,6 +582,7 @@ class FlawAffectsTrackersField(serializers.Field):
 @extend_schema_serializer(deprecate_fields=["mitigated_by"])
 class FlawSerializer(
     ACLMixinSerializer,
+    BugzillaSyncMixinSerializer,
     TrackingMixinSerializer,
     WorkflowModelSerializer,
     IncludeExcludeFieldsMixin,
@@ -662,6 +728,7 @@ class FlawSerializer(
                 "package_versions",
             ]
             + ACLMixinSerializer.Meta.fields
+            + BugzillaSyncMixinSerializer.Meta.fields
             + TrackingMixinSerializer.Meta.fields
             + WorkflowModelSerializer.Meta.fields
         )