+++ osidb/tests/test_endpoints.py
@@ -4,6 +4,7 @@ from typing import Set, Union
 
 import pytest
 from django.conf import settings
+from django.contrib.auth.models import Group, User
 from django.utils import timezone
 from django.utils.timezone import datetime
 from freezegun import freeze_time
@@ -1057,6 +1058,7 @@ class TestEndpoints(object):
             "unembargo_dt": "2000-1-1T22:03:26.065Z",
             "cvss3": "3.7/CVSS:3.0/AV:N/AC:H/PR:N/UI:N/S:U/C:N/I:L/A:N",
             "embargoed": False,
+            "bz_api_key": "SECRET",
         }
         response = auth_client.post(f"{test_api_uri}/flaws", flaw_data, format="json")
         assert response.status_code == 201
@@ -1083,6 +1085,7 @@ class TestEndpoints(object):
             "unembargo_dt": "2000-1-1T22:03:26.065Z",
             "cvss3": "3.7/CVSS:3.0/AV:N/AC:H/PR:N/UI:N/S:U/C:N/I:L/A:N",
             "embargoed": False,
+            "bz_api_key": "SECRET",
         }
         response = auth_client.post(f"{test_api_uri}/flaws", flaw_data, format="json")
         assert response.status_code == 201
@@ -1129,6 +1132,7 @@ class TestEndpoints(object):
                 "resolution": flaw.resolution,
                 "impact": flaw.impact,
                 "embargoed": False,
+                "bz_api_key": "SECRET",
             },
             format="json",
         )
@@ -1217,6 +1221,7 @@ class TestEndpoints(object):
             "ps_module": "rhacm-2",
             "ps_component": "curl",
             "embargoed": False,
+            "bz_api_key": "SECRET",
         }
         response = auth_client.post(
             f"{test_api_uri}/affects", affect_data, format="json"
@@ -1245,6 +1250,7 @@ class TestEndpoints(object):
             {
                 **original_body,
                 "ps_module": f"different {affect.ps_module}",
+                "bz_api_key": "SECRET",
             },
             format="json",
         )
@@ -1256,12 +1262,17 @@ class TestEndpoints(object):
         """
         Test the deletion of Affect records via a REST API DELETE request.
         """
-        affect = AffectFactory()
+        flaw = FlawFactory()
+        # an extra affect needs to be created as otherwise
+        # we would endup with an invalid affect-less flaw
+        AffectFactory(flaw=flaw)
+        affect = AffectFactory(flaw=flaw)
+
         affect_url = f"{test_api_uri}/affects/{affect.uuid}"
         response = auth_client.get(affect_url)
         assert response.status_code == 200
 
-        response = auth_client.delete(affect_url)
+        response = auth_client.delete(affect_url, data={"bz_api_key": "SECRET"})
         assert response.status_code == 204
 
         response = auth_client.get(affect_url)
@@ -1344,6 +1355,11 @@ class TestEndpointsACLs:
         """
         test proper embargo status and ACLs when creating a flaw by sending a POST request
         """
+        # we have to provide embargo write access first
+        group = Group(name="data-topsecret-write")
+        group.save()
+        User.objects.get(username="testuser").groups.add(group)
+
         flaw_data = {
             "title": "EMBARGOED Foo" if embargoed else "Foo",
             "description": "test",
@@ -1352,6 +1368,7 @@ class TestEndpointsACLs:
             "unembargo_dt": None if embargoed else "2000-1-1T22:03:26.065Z",
             "cvss3": "3.7/CVSS:3.0/AV:N/AC:H/PR:N/UI:N/S:U/C:N/I:L/A:N",
             "embargoed": embargoed,
+            "bz_api_key": "SECRET",
         }
         response = auth_client.post(f"{test_api_uri}/flaws", flaw_data, format="json")
         assert response.status_code == 201
@@ -1380,6 +1397,11 @@ class TestEndpointsACLs:
         test proper embargo status and ACLs when updating a flaw by sending a PUT request
         while the embargo status and ACLs itself are not being changed
         """
+        # we have to provide embargo write access first
+        group = Group(name="data-topsecret-write")
+        group.save()
+        User.objects.get(username="testuser").groups.add(group)
+
         flaw = FlawFactory(embargoed=embargoed)
         AffectFactory(flaw=flaw)
 
@@ -1394,6 +1416,7 @@ class TestEndpointsACLs:
                 "title": f"{flaw.title} appended test title",
                 "description": flaw.description,
                 "embargoed": embargoed,
+                "bz_api_key": "SECRET",
             },
             format="json",
         )
@@ -1427,6 +1450,7 @@ class TestEndpointsACLs:
                     "title": flaw.title.replace("EMBARGOED", "").strip(),
                     "description": flaw.description,
                     "embargoed": False,
+                    "bz_api_key": "SECRET",
                 },
                 format="json",
             )
@@ -1435,3 +1459,102 @@ class TestEndpointsACLs:
         body = response.json()
         assert body["embargoed"] is False
         assert Flaw.objects.first().embargoed is False
+
+    def test_flaw_create_not_member(self, auth_client, test_api_uri):
+        """
+        test that creating a Flaw is rejected when the ACL contains a group the user is not a member of
+        """
+        # restrict the user to the public read and write access
+        User.objects.get(username="testuser").groups.exclude(
+            name__in=["data-prodsec", "data-prodsec-write"]
+        ).delete()
+
+        flaw_data = {
+            "title": "EMBARGOED Foo",
+            "description": "test",
+            "reported_dt": "2022-11-22T15:55:22.830Z",
+            "unembargo_dt": None,
+            "cvss3": "3.7/CVSS:3.0/AV:N/AC:H/PR:N/UI:N/S:U/C:N/I:L/A:N",
+            "embargoed": True,
+            "bz_api_key": "SECRET",
+        }
+        response = auth_client.post(f"{test_api_uri}/flaws", flaw_data, format="json")
+        assert response.status_code == 400
+        assert (
+            "Cannot provide access for the LDAP group without being a member: data-topsecret"
+            in str(response.content)
+        )
+
+    def test_flaw_update_not_member(self, auth_client, test_api_uri):
+        """
+        test that updating a Flaw is rejected when the ACL contains a group the user is not a member of
+        """
+        flaw = FlawFactory(embargoed=False)
+        AffectFactory(flaw=flaw)
+
+        # restrict the user to the public read-only access
+        User.objects.get(username="testuser").groups.exclude(
+            name="data-prodsec"
+        ).delete()
+
+        response = auth_client.get(f"{test_api_uri}/flaws/{flaw.uuid}")
+        assert response.status_code == 200
+
+        response = auth_client.put(
+            f"{test_api_uri}/flaws/{flaw.uuid}",
+            {
+                "title": f"{flaw.title} appended test title",
+                "description": flaw.description,
+                "embargoed": False,
+                "bz_api_key": "SECRET",
+            },
+            format="json",
+        )
+        assert response.status_code == 400
+        assert (
+            "Cannot provide access for the LDAP group without being a member: data-prodsec-write"
+            in str(response.content)
+        )
+
+
+class TestEndpointsBZAPIKey:
+    """
+    Bugzilla API key specific tests
+    """
+
+    def test_flaw_create_no_bz_api_key(self, auth_client, test_api_uri):
+        """
+        test that creating a Flaw is rejected when no Bugzilla API key is provided
+        """
+        flaw_data = {
+            "title": "Foo",
+            "description": "test",
+            "reported_dt": "2022-11-22T15:55:22.830Z",
+            "unembargo_dt": "2000-1-1T22:03:26.065Z",
+            "cvss3": "3.7/CVSS:3.0/AV:N/AC:H/PR:N/UI:N/S:U/C:N/I:L/A:N",
+            "embargoed": False,
+        }
+        response = auth_client.post(f"{test_api_uri}/flaws", flaw_data, format="json")
+        assert response.status_code == 400
+        assert '"bz_api_key":["This field is required."]' in str(response.content)
+
+    def test_flaw_update_no_bz_api_key(self, auth_client, test_api_uri):
+        """
+        test that updating a Flaw is rejected when no Bugzilla API key is provided
+        """
+        flaw = FlawFactory(embargoed=False)
+        AffectFactory(flaw=flaw)
+        response = auth_client.get(f"{test_api_uri}/flaws/{flaw.uuid}")
+        assert response.status_code == 200
+
+        response = auth_client.put(
+            f"{test_api_uri}/flaws/{flaw.uuid}",
+            {
+                "title": f"{flaw.title} appended test title",
+                "description": flaw.description,
+                "embargoed": False,
+            },
+            format="json",
+        )
+        assert response.status_code == 400
+        assert '"bz_api_key":["This field is required."]' in str(response.content)